+++ src/Contracts/OrderStatus.php
@@ -13,4 +13,17 @@ namespace Vanilo\Order\Contracts;
 
 interface OrderStatus
 {
+    /**
+     * Returns whether the status represents an open state
+     *
+     * @return boolean
+     */
+    public function isOpen(): bool;
+
+    /**
+     * Returns an array of statuses that represent an open order state
+     *
+     * @return array
+     */
+    public static function getOpenStatuses(): array;
 }